SIM-380 UPDATE moved functionality to apply rules from utils to the rule
RefactoringRules act more as a service now and they do no longer hold their own working copies. 
The generateDocumentChanges method was moved to the new location where working copies are saved: which is RefactoringState. 

However, this meant that the OrganiseImports and CodeFormatter rules could no longer override generateDocumentChanges. 

For that reason, the applyRule method has been moved to the rule itself. Where overriding is possible.