SIM-1826 UPDATE check functional interface type compatibilities
Say current functional interface is the type of the lambda expression and the expected functional interface is the type of the expected context. The parameters of the current functional interface method should be supertypes of the expected functional interface, and the return type of the current functional interface shouild be a subtype of the expected context one.